consolidate to single index file
