--- types/jsmediatags/build2/ID3v1TagReader.d.ts
@@ -1,17 +1,0 @@
-import {
-    LoadCallbackType,
-    ByteRange,
-    TagType
-} from '../types';
-import MediaFileReader from './MediaFileReader';
-import MediaTagReader from './MediaTagReader';
-
-export default class ID3v1TagReader extends MediaTagReader {
-    static getTagIdentifierByteRange(): ByteRange;
-
-    static canReadTagFormat(tagIdentifier: number[]): boolean;
-
-    _loadData(mediaFileReader: MediaFileReader, callbacks: LoadCallbackType): void;
-
-    _parseData(data: MediaFileReader, tags?: string[]): TagType;
-}
